+++ libs/compiler/tree.h
@@ -1,0 +1,35 @@
+/*
+ *	Copyright 2020 Andrey Terekhov, Victor Y. Fadeev, Dmitrii Davladov
+ *
+ *	Licensed under the Apache License, Version 2.0 (the "License");
+ *	you may not use this file except in compliance with the License.
+ *	You may obtain a copy of the License at
+ *
+ *		http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *	Unless required by applicable law or agreed to in writing, software
+ *	distributed under the License is distributed on an "AS IS" BASIS,
+ *	WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *	See the License for the specific language governing permissions and
+ *	limitations under the License.
+ */
+
+#pragma once
+
+#include "syntax.h"
+
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+/**
+ *	Test tree building
+ *
+ *	@param	sx	Syntax structure
+ */
+void tree_test(const syntax *const sx);
+
+#ifdef __cplusplus
+} /* extern "C" */
+#endif
